refactor(ui): split create.tsx god object into create/ component modules
