+++ scripts/.god-object-baseline.txt
@@ -1,7 +1,6 @@
 # God-object baseline (LOC ceilings). Files here may only SHRINK.
 # Regenerate after a real split: scripts/check-file-sizes.sh --update
 # Doctrine: CLAUDE.md 'Module Size Discipline' — split by responsibility toward <600 LOC.
-2966 ui/src/routes/create.tsx
 1281 ui/src/routes/sandboxes.$id.tsx
 1106 ui/src/routes/instances.$id.tsx
 1095 ai-agent-sandbox-blueprint-lib/src/workflows.rs
